requirements.txt
geopandas~=0.12.2
matplotlib~=3.7.1
googlemaps~=4.10.0
python-dotenv~=0.21.0
flask~=2.2.2
pymongo~=3.12.0
ortools~=9.6.2534
aiohttp~=3.8.3
gunicorn==20.1.0
